terminar ejercicio 54

+++ Programacion2/src/ej47/Articulo.java
@@ -1,8 +1,10 @@
 package ej47;
 
+import java.math.BigDecimal;
+
 public abstract class Articulo {
 	private String descripcion;
-	private Double precio;
+	private BigDecimal precio;
 	
 	public String getDescripcion() {
 		return descripcion;
@@ -10,10 +12,10 @@ public abstract class Articulo {
 	public void setDescripcion(String descripcion) {
 		this.descripcion = descripcion;
 	}
-	public Double getPrecio() {
+	public BigDecimal getPrecio() {
 		return precio;
 	}
-	public void setPrecio(Double precio) {
+	public void setPrecio(BigDecimal precio) {
 		this.precio = precio;
 	}
 	@Override
